requirements.txt
# ============================================================================
# Django Gesture Presentation System - Requirements
# ============================================================================
# 
# Installation:
#   pip install -r requirements.txt
#
# For development with all tools:
#   pip install -r requirements.txt
#
# ============================================================================

# ============================================================================
# Core Django Framework
# ============================================================================
Django==4.2.27
asgiref==3.11.0
sqlparse==0.5.4
pytz==2025.2
tzdata==2025.2

# ============================================================================
# Django Extensions & Plugins
# ============================================================================
# REST API framework
djangorestframework==3.14.0

# CORS support for cross-origin requests
django-cors-headers==4.3.1

# Environment variable management
django-environ==0.11.2
python-dotenv==1.0.1

# ============================================================================
# Database Drivers
# ============================================================================
# PostgreSQL driver (optional - uncomment if using PostgreSQL)
psycopg2-binary==2.9.9

# ============================================================================
# ASGI & WebSocket Support
# ============================================================================
# ASGI server for async Django support
uvicorn==0.29.0
uvloop==0.19.0
websockets==12.0

# Django Channels for WebSocket support (optional - currently commented in asgi.py)
# channels==4.0.0
# channels-redis==4.2.0  # If using Redis as channel layer

# ============================================================================
# Image & Data Processing (if needed for future features)
# ============================================================================
# Image processing
Pillow==10.3.0

# Data analysis (optional - currently not directly used in code)
numpy==1.26.4
pandas==2.2.1

# ============================================================================
# Development Tools
# ============================================================================
# Code formatting
black==24.1.1
pathspec==0.12.1
mypy_extensions==1.1.0
click==8.1.8
platformdirs==4.4.0

# Code linting
flake8==7.0.0
pycodestyle==2.11.1
pyflakes==3.2.0
mccabe==0.7.0

# ============================================================================
# Testing Tools
# ============================================================================
pytest==7.4.4
pytest-django==4.8.0
iniconfig==2.1.0
pluggy==1.6.0
tomli==2.3.0

# ============================================================================
# Common Dependencies
# ============================================================================
# Standard library extensions
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.17.0
exceptiongroup==1.3.1

# HTTP/Network libraries
h11==0.16.0
packaging==25.0
